requirements.txt
torch
numpy
scipy
tqdm
matplotlib
pandas
numpy-stl
trimesh
stl
scikit-learn
psutil
pyyaml
deap
sympy
pytest
pytest-cov
plotly
seaborn
memory-profiler
joblib
torch-audio
torchvision
sphinx
sphinx-rtd-theme
jupyter
ipykernel
black
flake8
mypy
